Moved NetworkEdge and NetworkNode classes to separate files
